producer/requirements.txt
kafka-python-ng==2.2.3
Faker==30.8.2
